% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemodats.R
\docType{data}
\name{chemodats}
\alias{chemodats}
\title{This is a small network  included in the GraphsTutorial package
The data are a dataframe downloaded as a txt file from
the STRING data base.}
\format{
An dataframe with 20 rows corresponding to edges
between genes in small genetic network, the first two columns are
the nodes that define the edges, the other variables perain to
the choices and weights of the edges as generated by STRING.
}
\usage{
data(chemodats)
}
\description{
This is a small network  included in the GraphsTutorial package
The data are a dataframe downloaded as a txt file from
the STRING data base.
}
\references{
\url{https://www.huber.embl.de/msmb/Chap-Graphs.html#minimum-spanning-trees} and
\url{https://bios221.stanford.edu/book/Chap-Graphs.html#minimum-spanning-trees}
}
\author{
Susan Holmes \email{sp.holmes@gmail.com}
}
\keyword{datasets}
