% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2009c.R
\docType{data}
\name{dist2009c}
\alias{dist2009c}
\title{This is the dist2009c data  included in the GraphsTutorial package
The distance is a dist object obtained by using the
DNA distance between strains of HIV, the strain samples
come from HIV patients the Labels attribute contains the
country of origin.}
\format{
An object of class dist between 512 samples.
}
\usage{
data(dist2009c)
}
\description{
This is the dist2009c data  included in the GraphsTutorial package
The distance is a dist object obtained by using the
DNA distance between strains of HIV, the strain samples
come from HIV patients the Labels attribute contains the
country of origin.
}
\references{
\url{https://www.huber.embl.de/msmb/Chap-Graphs.html#minimum-spanning-trees} and
\url{https://bios221.stanford.edu/book/Chap-Graphs.html#minimum-spanning-trees}
}
\author{
Susan Holmes \email{sp.holmes@gmail.com}
}
\keyword{datasets}
