% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micephyloseq.R
\docType{data}
\name{micephyloseq}
\alias{micephyloseq}
\title{This is the mice phyloseq object data  included in the GraphsTutorial package
The data are a multicomponent phyloseq object obtained
from several samples from different mice belonging to two litters.}
\format{
An object of class phyloseq with 344 different samples
and 389 taxa, there are 16 sample variables of which the most
important are SampleID and family_relationship.
}
\usage{
data(micephyloseq)
}
\description{
This is the mice phyloseq object data  included in the GraphsTutorial package
The data are a multicomponent phyloseq object obtained
from several samples from different mice belonging to two litters.
}
\references{
\url{https://www.huber.embl.de/msmb/Chap-Graphs.html#minimum-spanning-trees} and
\url{https://bios221.stanford.edu/book/Chap-Graphs.html#minimum-spanning-trees}
}
\author{
Susan Holmes \email{sp.holmes@gmail.com}
}
\keyword{datasets}
